// santa-front/src/types/FormGroupCreationActions.enum.tsx
const enum FormGroupCreationActions {
	ADD_USER = 'ADD_USER',
	REMOVE_USER = 'REMOVE_USER',
	UPDATE_MAIN_USER_ID = 'UPDATE_MAIN_USER_ID',
	UPDATE_USER = 'UPDATE_USER',
	UPDATE_USERS = 'UPDATE_USERS',
	UPDATE_GROUP_NAME = 'UPDATE_GROUP_NAME',
	UPDATE_STEP = 'UPDATE_STEP',
	UPDATE_USER_EXCLUSION = 'UPDATE_USER_EXCLUSION',
	RESET = 'RESET',
	UPDATE_DUE_DATE = 'UPDATE_DUE_DATE'
};

export default FormGroupCreationActions;